Make minor changes: 1. Add Lombok to build.gradle. 2. Remove version info in build.gradle. 3. Change pojo to domain. 4. Update Bootstrap from 4 to 5. 5. Change field injection to constructor injection.
