+++ src/main/java/netty/springboot/multipersonchatroom/controller/PageController.java
@@ -1,17 +1,17 @@
 package netty.springboot.multipersonchatroom.controller;
 
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
 public class PageController {
 
-    @RequestMapping("/")
+    @GetMapping("/")
     public String index() {
         return "index";
     }
 
-    @RequestMapping("/chat")
+    @GetMapping("/chat")
     public String chat() {
         return "chat";
     }
